# Conflicts:
#	src/core.rs
#	tests/simulation.rs

+++ src/entity/my_identity.rs
@@ -8,7 +8,7 @@ pub struct Model {
 	#[sea_orm(primary_key, auto_increment = false)]
 	pub label: String,
 	#[sea_orm(unique)]
-	pub identity_id: i32,
+	pub identity_id: i64,
 	#[sea_orm(column_type = "Binary(BlobSize::Blob(None))")]
 	pub private_key: Vec<u8>,
 }
